pyproject.toml
[tool.poetry]
name = "asset-tracking"
version = "0.1.0"
description = "Enterprise asset tracking by hostname for rouge device detection."
authors = ["Sean McFeely <mcfeelynaes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
argcomplete = "^2.0.0"
coloredlogs = "^15.0.0"
python-dateutil = "^2.8.0"
SQLAlchemy = "^1.3.0"
pydantic = "^1.8.0"
SQLAlchemy-Utils = "^0.38.3"
pg8000 = "^1.29.1"
configparser = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.971"
flake8 = "^4.0.1"
black = "^22.6.0"
types-python-dateutil = "^2.8.19"
pydocstyle = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.poetry.scripts]
asset-tracker = "asset_tracking.cli:main"